This is a huge set of changes
UI Related: DPI, to translations added support for 30+ languages, new icons, new skins, and various other improvements and bug fixes.
Engine Related: New OPUS codec, new WASAPI output, new ReplayGain support, new DiscordRPC support, new TaskbarIntegration support, and various other improvements and bug fixes.
Codecs: Made FFMPEG as optional when building with CMake, added support for building with minimp4 and fdk-aac, improved flac support etc, and various other improvements and bug fixes.

+++ CPString.c
@@ -150,7 +150,7 @@ unsigned int STR_AllocSetStringW(WCHAR** ppwcDest, const WCHAR* pwcSource, const
 	{
 		unsigned int uStringLength;
 		
-		uStringLength = (wcslen(pwcSource) + 1) * sizeof(WCHAR);
+		uStringLength = (unsigned int)((wcslen(pwcSource) + 1) * sizeof(WCHAR));
 		*ppwcDest = CALLOC_TYPE(WCHAR, wcslen(pwcSource) + 1);
 		
 		if (!*ppwcDest)
